Replace let keyword with const. Add support for const functions and classes

+++ src/debug.c
@@ -63,7 +63,7 @@ static size_t disassemble_op_constant_long(const char *name, const Chunk *chunk,
 static size_t disassemble_op_jump(const char *name, const int sign, const Chunk *chunk, const size_t offset) {
     uint16_t jump = chunk->code[offset + 1];
     jump |= (uint16_t)chunk->code[offset + 2] << 8;
-    printf("%-16s %4zu -> %zu\n", name, offset, offset + 3 + sign * jump);
+    printf("%-16s %4zu -> %zu\n", name, offset, offset + 3 + (size_t)sign * jump);
     return offset + 3;
 }
